LayoutTests/fast/dom/timer-throttling-hidden-page-expected.txt

Bug 98474: Throttle DOM timers on hidden pages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS timerIntervalWhilePageVisible is within 20 of 10
PASS timerIntervalWhilePageVisible is within 20 of 10
PASS timerIntervalWhilePageVisible is within 20 of 10
PASS timerIntervalWhilePageVisible is within 20 of 10
PASS timerIntervalWhilePageVisible is within 20 of 10
PASS timerIntervalWhilePageHidden > (timeoutInterval + tolerance) is true
PASS timerIntervalWhilePageVisible is within 20 of 10
PASS successfullyParsed is true

TEST COMPLETE
This test measures the time taken to fire a 10ms DOM Timer when the page visibility is set to "visible", "hidden" and then back to "visible". Due to hidden page timer throttling, the timer should fire close to s when page is hidden. And it should fire close to 10ms, when the page is visible.
